Changed int to short, added array generation

+++ Sorting Test/Drawer.cpp
@@ -4,7 +4,7 @@ Drawer::~Drawer()
 {
 }
 
-void Drawer::DrawArray(int* data, int size)
+void Drawer::DrawArray(short* data, short size)
 {
 	Chart->Series[0]->Points->Clear();
 	for (size_t i = 0; i < size; i++)
